+++ src/schematools/validation.py
@@ -29,9 +29,9 @@ from typing import Callable, Iterator, Optional, Set, cast
 from urllib.parse import urlparse
 
 from schematools import MAX_TABLE_NAME_LENGTH
-from schematools.exceptions import SchemaObjectNotFound
+from schematools.exceptions import DatasetFieldNotFound, SchemaObjectNotFound
 from schematools.naming import to_snake_case, toCamelCase
-from schematools.types import DatasetSchema, SemVer, TableVersions
+from schematools.types import DatasetSchema, TableVersions
 
 
 @dataclass(frozen=True)
@@ -43,21 +43,6 @@ class ValidationError:
 
     def __str__(self) -> str:
         return f"[{self.validator_name}] {self.message}"
-
-
-class ValidationException(Exception):
-    """Raised when validation fails to execute.
-
-    .. note::
-
-       This is not for validation errors. See :class:`ValidationError` instead.
-    """
-
-    message: str
-
-    def __init__(self, message: str) -> None:  # noqa: D107
-        super().__init__(message)
-        self.message = message
 
 
 _all: list[tuple[str, Callable[[DatasetSchema, str | None], Iterator[str]]]] = []
@@ -176,6 +161,19 @@ def _id_type(dataset: DatasetSchema) -> Iterator[str]:
                     )
             except SchemaObjectNotFound as e:
                 yield f"{ident!r} listed in identifier list {table.identifier}, but: {e}"
+
+
+@_register_validator("'id' field in the presence of composite primary key")
+def _id_and_composite(dataset: DatasetSchema) -> Iterator[str]:
+    for table in dataset.tables:
+        if len(table.identifier) < 2:
+            continue
+        try:
+            table.get_field_by_id("id")
+        except DatasetFieldNotFound:
+            pass
+        else:
+            yield f"table {table.qualified_id!r} has a field called 'id'"
 
 
 @_register_validator("PostgreSQL identifier length")